Python/INS: clean up names
 - PyINS is the pure python implementation
 - CINS is a wrapper to the C implementation
