+++ python/ins/__init__.py
@@ -3,6 +3,6 @@ from __future__ import absolute_import # Use 3.x import model.
 #-------------------------------------------------------------------------------
 __all__ = ()
 
-from . import ins14
+from . import cins
 from . import pyins
